requirements.txt
numpy>=1.24.0
openai>=1.12.0
python-dotenv>=1.0.0
